review: move open_for_read to recall.rs (resolves any store, not just remotes)
open_for_read handles local stores too, so it doesn't belong in the
Hub-remote file. Its only callers are the read verbs in recall.rs, so it
— plus ReadOutcome and the is_missing_dataset detector (and its tests) —
moves there and becomes private. hub.rs keeps the genuinely remote-state
pieces it calls into: Reachability and the shared missing_remote /
empty_remote messages (still used by push and use).
